Code cleanup

+++ app/src/main/java/com/sukie2/android/currencyconverter/extentions/ConverterExtentions.kt
@@ -1,7 +1,5 @@
 package com.sukie2.android.currencyconverter.extentions
 
-import java.text.DecimalFormat
-import java.text.DecimalFormatSymbols
 import java.util.*
 
 
@@ -9,7 +7,7 @@ import java.util.*
  * Format the float to string
  * Locale safe
  */
-fun Double.format() : String = String.format(Locale.getDefault(), "%.2f", this)
+fun Float.format() : String = String.format(Locale.getDefault(), "%.2f", this)
 
 /**
  * Returns empty ("") if the string object is null
